data with reference double in C, better than python?

+++ exp_rapide/exp_rapide.h
@@ -1,5 +1,5 @@
-typedef float TYPE;
-//typedef double TYPE;
+//typedef float TYPE;
+typedef double TYPE;
 
 
 TYPE exp_rapide_rec(int n, TYPE x);
